shill: Add lease_name field at DHCPCDControllerInterface::Options
DHCPCDControllerInterface::Options is the duplicatation of the current
DHCPController::Options. Originally we thought the lease_name field is
not used anymore, so we didn't copy the field. However, the WiFi class
still uses the field.

This CL adds the field back to DHCPCDControllerInterface::Options, and
implements the related logic at LegacyDHCPCDController.

BUG=b:342068333
TEST=FEATURES=test emerge-brya shill

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5610632
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>

+++ shill/network/legacy_dhcpcd/legacy_dhcpcd_controller.cc
@@ -88,6 +88,13 @@ std::vector<std::string> GetDhcpcdFlags(
   return flags;
 }
 
+// Returns true if the lease file is ephermeral, which means the lease file
+// should be deleted during cleanup.
+bool IsEphemeralLease(const DHCPCDControllerInterface::Options& options,
+                      std::string_view interface) {
+  return options.lease_name.empty() || options.lease_name == interface;
+}
+
 }  // namespace
 
 LegacyDHCPCDController::LegacyDHCPCDController(
@@ -174,7 +181,11 @@ LegacyDHCPCDControllerFactory::Create(
     const DHCPCDControllerInterface::Options& options,
     DHCPCDControllerInterface::EventHandler* handler) {
   std::vector<std::string> args = GetDhcpcdFlags(technology, options);
-  args.push_back(std::string(interface));
+  if (IsEphemeralLease(options, interface)) {
+    args.push_back(std::string(interface));
+  } else {
+    args.push_back(base::StrCat({interface, "=", options.lease_name}));
+  }
 
   net_base::ProcessManager::MinijailOptions minijail_options;
   minijail_options.user = kDHCPCDUser;
@@ -232,8 +243,10 @@ void LegacyDHCPCDControllerFactory::CleanUpDhcpcd(
   }
 
   // Clean up the lease file and pid file.
-  brillo::DeleteFile(root_.Append(
-      base::StringPrintf(kDHCPCDPathFormatLease, interface.c_str())));
+  if (IsEphemeralLease(options, interface)) {
+    brillo::DeleteFile(root_.Append(
+        base::StringPrintf(kDHCPCDPathFormatLease, interface.c_str())));
+  }
   brillo::DeleteFile(root_.Append(
       base::StringPrintf(kDHCPCDPathFormatPID, interface.c_str())));
 }
